add additional form options for divident calc
